Output_parser/requirements.txt
# LangChain Core
langchain
langchain-core


# Groq Integration
langchain-groq
groq

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
pydantic
pydantic[email]